Move animation to separated module
