+++ src/reanimated2/index.ts
@@ -1,6 +1,6 @@
 export * from './core';
 export * from './Hooks';
-export * from './animations';
+export * from './animation';
 export * from './interpolation';
 export * from './Easing';
 export * from './NativeMethods';
