+++ src/reanimated2/layoutReanimation/defaultAnimations/Bounce.ts
@@ -3,7 +3,7 @@ import {
   EntryExitAnimationBuild,
 } from '../animationBuilder/commonTypes';
 import { BaseBounceAnimationBuilder } from '../animationBuilder/BaseBounceAnimationBuilder';
-import { withSequence, withTiming } from '../../animations';
+import { withSequence, withTiming } from '../../animation';
 import { Dimensions } from 'react-native';
 
 const { width, height } = Dimensions.get('window');
